// src/generated/zod.gen.ts
export const zAccount = z.object({
  id: z.uuid(),
  platform_id: z.string(),
  parent_id: z.uuid().nullish(),
  account_type: z.enum([
    'accounts_payable',
    'accounts_receivable',
    'bank',
    'fixed_asset',
    'other_asset',
    'other_current_asset',
    'liability',
    'equity',
    'expense',
    'other_expense',
    'income',
    'other_income',
    'credit_card',
    'cost_of_goods_sold',
    'other_current_liability',
    'long_term_liability',
    'non_posting',
    'unknown',
  ]),
  category: z.enum([
    'asset',
    'expense',
    'equity',
    'liability',
    'income',
    'nonposting',
    'unknown',
  ]),
  status: z.enum(['active', 'inactive', 'pending']),
  balance: z.number().nullable(),
  currency_code: z.enum([
    'AED',
    'AFN',
    'ALL',
    'AMD',
    'ANG',
    'AOA',
    'ARS',
    'AUD',
    'AWG',
    'AZN',
    'BAM',
    'BBD',
    'BDT',
    'BGN',
    'BHD',
    'BIF',
    'BMD',
    'BND',
    'BOB',
    'BOV',
    'BRL',
    'BSD',
    'BTN',
    'BWP',
    'BYR',
    'BZD',
    'CAD',
    'CDF',
    'CHE',
    'CHF',
    'CHW',
    'CLF',
    'CLP',
    'CNY',
    'COP',
    'COU',
    'CRC',
    'CUC',
    'CUP',
    'CVE',
    'CZK',
    'DJF',
    'DKK',
    'DOP',
    'DZD',
    'EGP',
    'ERN',
    'ETB',
    'EUR',
    'FJD',
    'FKP',
    'GBP',
    'GEL',
    'GHS',
    'GIP',
    'GMD',
    'GNF',
    'GTQ',
    'GYD',
    'HKD',
    'HNL',
    'HRK',
    'HTG',
    'HUF',
    'IDR',
    'ILS',
    'INR',
    'IQD',
    'IRR',
    'ISK',
    'JMD',
    'JOD',
    'JPY',
    'KES',
    'KGS',
    'KHR',
    'KMF',
    'KPW',
    'KRW',
    'KWD',
    'KYD',
    'KZT',
    'LAK',
    'LBP',
    'LKR',
    'LRD',
    'LSL',
    'LTL',
    'LVL',
    'LYD',
    'MAD',
    'MDL',
    'MGA',
    'MKD',
    'MMK',
    'MNT',
    'MOP',
    'MRO',
    'MUR',
    'MVR',
    'MWK',
    'MXN',
    'MXV',
    'MYR',
    'MZN',
    'NAD',
    'NGN',
    'NIO',
    'NOK',
    'NPR',
    'NZD',
    'OMR',
    'PAB',
    'PEN',
    'PGK',
    'PHP',
    'PKR',
    'PLN',
    'PYG',
    'QAR',
    'RON',
    'RSD',
    'RUB',
    'RWF',
    'SAR',
    'SBD',
    'SCR',
    'SDG',
    'SEK',
    'SGD',
    'SHP',
    'SLL',
    'SOS',
    'SRD',
    'SSP',
    'STD',
    'SVC',
    'SYP',
    'SZL',
    'THB',
    'TJS',
    'TMT',
    'TND',
    'TOP',
    'TRY',
    'TTD',
    'TWD',
    'TZS',
    'UAH',
    'UGX',
    'USD',
    'USN',
    'USS',
    'UYI',
    'UYU',
    'UYW',
    'UZS',
    'VEF',
    'VND',
    'VUV',
    'WST',
    'XAF',
    'XAG',
    'XAU',
    'XBA',
    'XBB',
    'XBC',
    'XBD',
    'XCD',
    'XDR',
    'XFU',
    'XOF',
    'XPD',
    'XPF',
    'XPT',
    'XTS',
    'XXX',
    'YER',
    'ZAR',
    'ZMW',
    'ETH',
    'BTC',
    'null',
  ]),
  name: z.string().nullable(),
  nominal_code: z.string().nullable(),
  subsidiaries: z.array(
    z.object({
      id: z.uuid(),
      status: z.enum(['active', 'inactive']),
    }),
  ),
  created_at: z.string().nullable(),
  updated_at: z.string().nullable(),
  last_synced_at: z.string(),
  platform_url: z.string().nullable(),
  platform_data: z.record(z.string(), z.unknown()).optional(),
})

/**
 * The underlying platform.
 */
export const zPlatform = z.enum([
  'AMAZON',
  'EBAY',
  'ETSY',
  'FNAC',
  'WALMART',
  'SHOPEE',
  'LAZADA',
  'MERCADOLIBRE',
  'ALIBABA',
  'SHOPLINE',
  'PRESTASHOP',
  'SHOPIFY',
  'MAGENTO',
  'WOO_COMMERCE',
  'SQUARESPACE',
  'WIX',
  'SHOPER',
  'SHOPLAZZA',
  'BIG_COMMERCE',
  'SALLA',
  'ZID',
  'COMMERCELAYER',
  'WEBFLOW',
  'GUMROAD',
  'ECWID',
  'PAYPAL',
  'SQUARE',
  'STRIPE',
  'CLOVER',
  'MOLLIE',
  'PAYNL',
  'AUTHORIZENET',
  'RECHARGE',
  'CHARGIFY',
  'CHARGEBEE',
  'RECURLY',
  'KASHFLOW',
  'SAGE_INTACCT',
  'SAGE_INTACCT_V2',
  'DYNAMICS365',
  'ZOHOBOOKS',
  'EXACTONLINE',
  'FREEAGENT',
  'QUICKBOOKS',
  'QUICKBOOKS_DESKTOP',
  'FRESHBOOKS',
  'XERO',
  'SAGE_BUSINESS_CLOUD',
  'SAGE_50',
  'NETSUITE',
  'WAVE',
  'DUALENTRY',
  'MONEYBIRD',
  'SAGE200CLOUD',
  'MX',
  'PLAID',
  'TELLER',
  'ODOO',
  'TALLY',
  'GOOGLE',
  'FACEBOOK',
  'TIKTOK',
  'INTUIT_BANK_FEEDS',
  'SAGE',
  'SAGE_BANK_FEEDS',
  'QUICKBOOKS_HRIS',
  'BAMBOO_HR',
  'GUSTO',
])

export const zConnection = z.object({
  id: z.uuid(),
  orgId: z.uuid(),
  platform: zPlatform,
})

export const zListAccountResponseWithConnection = z.object({
  accounts: z.array(zAccount),
  next_cursor: z.string().nullish(),
  connection: zConnection,
})
// ...
export const zAsyncResponsePayload = z.object({
  async_response: z
    .object({
      id: z.string(),
      response_url: z.string(),
      status: z.enum(['prequeued', 'pending', 'success', 'failure']),
    })
    .optional(),
})

export const zCreateAccountResponse = z.union([
  zAccountResponse,
  zAsyncResponsePayload,
])

export const zIdPathParam = z.object({
  id: z.string(),
})

export const zCommonGetParamsWithIdType = z.object({
  access_token: z.uuid(),
  id_type: z.enum(['rutter', 'platform']).optional(),
  force_fetch: z.enum(['true', 'false']).optional(),
})

export const zAccountResponseWithConnection = z.object({
  account: zAccount,
  connection: zConnection,
})
// ...
export const zCompanyInfo20240430 = z.object({
  id: z.string(),
  platform_id: z.string().nullable(),
  currency_code: z.enum([
    'AED',
    'AFN',
    'ALL',
    'AMD',
    'ANG',
    'AOA',
    'ARS',
    'AUD',
    'AWG',
    'AZN',
    'BAM',
    'BBD',
    'BDT',
    'BGN',
    'BHD',
    'BIF',
    'BMD',
    'BND',
    'BOB',
    'BOV',
    'BRL',
    'BSD',
    'BTN',
    'BWP',
    'BYR',
    'BZD',
    'CAD',
    'CDF',
    'CHE',
    'CHF',
    'CHW',
    'CLF',
    'CLP',
    'CNY',
    'COP',
    'COU',
    'CRC',
    'CUC',
    'CUP',
    'CVE',
    'CZK',
    'DJF',
    'DKK',
    'DOP',
    'DZD',
    'EGP',
    'ERN',
    'ETB',
    'EUR',
    'FJD',
    'FKP',
    'GBP',
    'GEL',
    'GHS',
    'GIP',
    'GMD',
    'GNF',
    'GTQ',
    'GYD',
    'HKD',
    'HNL',
    'HRK',
    'HTG',
    'HUF',
    'IDR',
    'ILS',
    'INR',
    'IQD',
    'IRR',
    'ISK',
    'JMD',
    'JOD',
    'JPY',
    'KES',
    'KGS',
    'KHR',
    'KMF',
    'KPW',
    'KRW',
    'KWD',
    'KYD',
    'KZT',
    'LAK',
    'LBP',
    'LKR',
    'LRD',
    'LSL',
    'LTL',
    'LVL',
    'LYD',
    'MAD',
    'MDL',
    'MGA',
    'MKD',
    'MMK',
    'MNT',
    'MOP',
    'MRO',
    'MUR',
    'MVR',
    'MWK',
    'MXN',
    'MXV',
    'MYR',
    'MZN',
    'NAD',
    'NGN',
    'NIO',
    'NOK',
    'NPR',
    'NZD',
    'OMR',
    'PAB',
    'PEN',
    'PGK',
    'PHP',
    'PKR',
    'PLN',
    'PYG',
    'QAR',
    'RON',
    'RSD',
    'RUB',
    'RWF',
    'SAR',
    'SBD',
    'SCR',
    'SDG',
    'SEK',
    'SGD',
    'SHP',
    'SLL',
    'SOS',
    'SRD',
    'SSP',
    'STD',
    'SVC',
    'SYP',
    'SZL',
    'THB',
    'TJS',
    'TMT',
    'TND',
    'TOP',
    'TRY',
    'TTD',
    'TWD',
    'TZS',
    'UAH',
    'UGX',
    'USD',
    'USN',
    'USS',
    'UYI',
    'UYU',
    'UYW',
    'UZS',
    'VEF',
    'VND',
    'VUV',
    'WST',
    'XAF',
    'XAG',
    'XAU',
    'XBA',
    'XBB',
    'XBC',
    'XBD',
    'XCD',
    'XDR',
    'XFU',
    'XOF',
    'XPD',
    'XPF',
    'XPT',
    'XTS',
    'XXX',
    'YER',
    'ZAR',
    'ZMW',
    'ETH',
    'BTC',
  ]),
  name: z.string(),
  addresses: z.array(z.string()),
  legal_name: z.string().nullable(),
  additional_fields: z
    .object({
      closing_date: z.string().nullish(),
    })
    .optional(),
  created_at: z.string().nullable(),
  updated_at: z.string().nullable(),
  last_synced_at: z.string(),
  platform_data: z.record(z.string(), z.unknown()).optional(),
})

export const zCompanyInfo20240430ResponseWithConnection = z.object({
  company_info: zCompanyInfo20240430,
  connection: zConnection,
})
// ...
export const zListInvoiceResponseWithConnection = z.object({
  invoices: z.array(zInvoice),
  next_cursor: z.string().nullish(),
  connection: zConnection,
})
// ...
export const zInvoiceResponse = z.object({
  invoice: zInvoice,
})

export const zCreateInvoiceResponse = z.union([
  zInvoiceResponse,
  zAsyncResponsePayload,
])

export const zInvoiceResponseWithConnection = z.object({
  invoice: zInvoice,
  connection: zConnection,
})